Search dialog works! (API < 11)
Problems: the theme is horrid. For some reason, the dialog is
all white on my phone. I'm guessing this is because I'm running
Android 4.2 but targeting a lower SDK, so it's a weird hybrid.

TODO: test this on a properly emulated 2.3 or something.

Another problem: live search doesn't work, and I doubt it ever will
because it would have to launch a new instance of the BookListActivity
for every change in character... oh well!